Trait gsk4::prelude::RenderNodeExt

pub trait RenderNodeExt: 'static {
    pub fn draw(&self, cr: &Context);

    pub fn get_bounds(&self) -> Rect;

    pub fn get_node_type(&self) -> RenderNodeType;

    pub fn serialize(&self) -> Option<Bytes>;

    pub fn write_to_file(&self, filename: &str) -> Result<(), Error>;
}

Trait containing all RenderNode methods.

Implementors

BlendNode, BlurNode, BorderNode, CairoNode, ClipNode, ColorMatrixNode, ColorNode, ConicGradientNode, ContainerNode, CrossFadeNode, DebugNode, GLShaderNode, InsetShadowNode, LinearGradientNode, OpacityNode, OutsetShadowNode, RadialGradientNode, RenderNode, RepeatNode, RepeatingLinearGradientNode, RepeatingRadialGradientNode, RoundedClipNode, ShadowNode, TextNode, TextureNode, TransformNode

Required methods

pub fn draw(&self, cr: &Context)

Draw the contents of self to the given cairo context.

Typically, you'll use this function to implement fallback rendering of GskRenderNodes on an intermediate Cairo context, instead of using the drawing context associated to a gdk::Surface's rendering buffer.

For advanced nodes that cannot be supported using Cairo, in particular for nodes doing 3D operations, this function may fail.

cr

cairo context to draw to

pub fn get_bounds(&self) -> Rect

Retrieves the boundaries of the self. The node will not draw outside of its boundaries.

bounds

return location for the boundaries

pub fn get_node_type(&self) -> RenderNodeType

Returns the type of the self.

Returns

the type of the RenderNode

pub fn serialize(&self) -> Option<Bytes>

Serializes the self for later deserialization via RenderNode::deserialize. No guarantees are made about the format used other than that the same version of GTK will be able to deserialize the result of a call to RenderNodeExt::serialize and RenderNode::deserialize will correctly reject files it cannot open that were created with previous versions of GTK.

The intended use of this functions is testing, benchmarking and debugging. The format is not meant as a permanent storage format.

Returns

a glib::Bytes representing the node.

pub fn write_to_file(&self, filename: &str) -> Result<(), Error>

This function is equivalent to calling RenderNodeExt::serialize followed by g_file_set_contents. See those two functions for details on the arguments.

It is mostly intended for use inside a debugger to quickly dump a render node to a file for later inspection.

filename

the file to save it to.

Returns

true if saving was successful

Loading content...