Trait gsk4::prelude::RendererExt

pub trait RendererExt: 'static {
    fn get_surface(&self) -> Option<Surface>;

    fn is_realized(&self) -> bool;

    fn realize(&self, surface: &Surface) -> Result<(), Error>;

    fn unrealize(&self);

    fn get_property_realized(&self) -> bool;

    fn connect_property_realized_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_surface_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;
}

Trait containing all Renderer methods.

Implementors

BroadwayRenderer, CairoRenderer, GLRenderer, Renderer, VulkanRenderer

Required methods

fn get_surface(&self) -> Option<Surface>

Retrieves the gdk::Surface set using RendererExt::realize. If the renderer has not been realized yet, None will be returned.

Returns

a gdk::Surface

fn is_realized(&self) -> bool

Checks whether the self is realized or not.

Returns

true if the Renderer was realized, and false otherwise

fn realize(&self, surface: &Surface) -> Result<(), Error>

Creates the resources needed by the self to render the scene graph.

surface

the gdk::Surface renderer will be used on

fn unrealize(&self)

Releases all the resources created by RendererExt::realize.

fn get_property_realized(&self) -> bool

fn connect_property_realized_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_surface_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

Loading content...