Struct gsk4::BlendNode

#[repr(transparent)]
pub struct BlendNode { /* private fields */ }

A render node applying a blending function between its two child nodes.

Implementations

impl BlendNode

pub fn as_ptr(&self) -> *mut GskBlendNode

Return the inner pointer to the underlying C value.

impl BlendNode

pub fn new(
bottom: impl AsRef<RenderNode>,
top: impl AsRef<RenderNode>,
blend_mode: BlendMode
) -> BlendNode

Creates a RenderNode that will use @blend_mode to blend the @top node onto the @bottom node.

bottom

The bottom node to be drawn

top

The node to be blended onto the @bottom node

blend_mode

The blend mode to use

Returns

A new RenderNode

pub fn blend_mode(&self) -> BlendMode

Retrieves the blend mode used by @self.

Returns

the blend mode

pub fn bottom_child(&self) -> RenderNode

Retrieves the bottom RenderNode child of the @self.

Returns

the bottom child node

pub fn top_child(&self) -> RenderNode

Retrieves the top RenderNode child of the @self.

Returns

the top child node