Struct gsk4::CairoNode

pub struct CairoNode(_);

A render node for a Cairo surface.

Implementations

impl CairoNode

pub fn new(bounds: &Rect) -> Self

Creates a RenderNode that will render a cairo surface into the area given by bounds.

You can draw to the cairo surface using draw_context().

bounds

the rectangle to render to

Returns

A new RenderNode

pub fn draw_context(&self) -> Option<Context>

Creates a Cairo context for drawing using the surface associated to the render node.

If no surface exists yet, a surface will be created optimized for rendering to renderer.

Returns

a Cairo context used for drawing; use cairo_destroy() when done drawing

pub fn surface(&self) -> Option<Surface>

Retrieves the Cairo surface used by the render node.

Returns

a Cairo surface