gsk4

Struct DebugNode

pub struct DebugNode { /* private fields */ }

A render node that emits a debugging message when drawing its child node.

GLib type: Shared boxed type with reference counted clone semantics.

Implementations

impl DebugNode

pub fn as_ptr(&self) -> *mut GskDebugNode

Return the inner pointer to the underlying C value.

pub unsafe fn from_glib_ptr_borrow(ptr: &*mut GskDebugNode) -> &Self

Borrows the underlying C value.

impl DebugNode

pub fn new(child: impl AsRef<RenderNode>, message: GString) -> DebugNode

Creates a RenderNode that will add debug information about the given @child.

Adding this node has no visual effect.

child

The child to add debug info for

message

The debug message

Returns

A new RenderNode

pub fn child(&self) -> RenderNode

Gets the child node that is getting drawn by the given @self.

Returns

the child RenderNode

pub fn message(&self) -> GString

Gets the debug message that was set on this node

Returns

The debug message