Struct gsk4::FillNode

#[repr(transparent)]
pub struct FillNode { /* private fields */ }

Available on crate feature v4_14 only.

A render node filling the area given by Path and FillRule with the child node.

Implementations

impl FillNode

pub fn as_ptr(&self) -> *mut GskFillNode

Return the inner pointer to the underlying C value.

pub unsafe fn from_glib_ptr_borrow<'a>( ptr: *const *const GskFillNode ) -> &'a Self

Borrows the underlying C value.

impl FillNode

pub fn new( child: impl AsRef<RenderNode>, path: &Path, fill_rule: FillRule ) -> FillNode

Creates a RenderNode that will fill the @child in the area given by @path and @fill_rule.

child

The node to fill the area with

path

The path describing the area to fill

fill_rule

The fill rule to use

Returns

A new RenderNode

pub fn child(&self) -> RenderNode

Gets the child node that is getting drawn by the given @self.

Returns

The child that is getting drawn

pub fn fill_rule(&self) -> FillRule

Retrieves the fill rule used to determine how the path is filled.

Returns

a FillRule

pub fn path(&self) -> Path

Retrieves the path used to describe the area filled with the contents of the @self.

Returns

a Path