Struct gsk4::GLShaderNode

pub struct GLShaderNode { /* private fields */ }

A render node using a GL shader when drawing its children nodes.

Implementations

impl GLShaderNode

pub fn as_ptr(&self) -> *mut GskGLShaderNode

Return the inner pointer to the underlying C value.

pub unsafe fn from_glib_ptr_borrow<'a>( ptr: *const *const GskGLShaderNode ) -> &'a Self

Borrows the underlying C value.

impl GLShaderNode

pub fn new( shader: &GLShader, bounds: &Rect, args: &Bytes, children: &[RenderNode] ) -> GLShaderNode

Creates a RenderNode that will render the given @shader into the area given by @bounds.

The @args is a block of data to use for uniform input, as per types and offsets defined by the @shader. Normally this is generated by Gsk::GLShader::format_args() or ShaderArgsBuilder.

See GLShader for details about how the shader should be written.

All the children will be rendered into textures (if they aren’t already GskTextureNodes, which will be used directly). These textures will be sent as input to the shader.

If the renderer doesn’t support GL shaders, or if there is any problem when compiling the shader, then the node will draw pink. You should use GLShader::compile() to ensure the @shader will work for the renderer before using it.

shader

the GLShader

bounds

the rectangle to render the shader into

args

Arguments for the uniforms

children

array of child nodes, these will be rendered to textures and used as input.

Returns

A new RenderNode

pub fn args(&self) -> Bytes

Gets args for the node.

Returns

A GBytes with the uniform arguments

pub fn n_children(&self) -> u32

Returns the number of children

Returns

The number of children

pub fn shader(&self) -> GLShader

Gets shader code for the node.

Returns

the GLShader shader

impl GLShaderNode

pub fn child(&self, idx: u32) -> RenderNode

Gets one of the children.

idx

the position of the child to get

Returns

the @idx’th child of @self