Struct gsk4::RenderNode

#[repr(transparent)]
pub struct RenderNode { /* private fields */ }

RenderNode is the basic block in a scene graph to be rendered using Renderer.

Each node has a parent, except the top-level node; each node may have children nodes.

Each node has an associated drawing surface, which has the size of the rectangle set when creating it.

Render nodes are meant to be transient; once they have been associated to a Renderer it’s safe to release any reference you have on them. All RenderNodes are immutable, you can only specify their properties during construction.

This is an Abstract Base Class, you cannot instantiate it.

Implementations

impl RenderNode

pub const NONE: Option<&'static RenderNode>

pub fn draw(&self, cr: &Context)

Draw the contents of @self to the given cairo context.

Typically, you’ll use this function to implement fallback rendering of RenderNodes on an intermediate Cairo context, instead of using the drawing context associated to a gdk::Surface’s rendering buffer.

For advanced nodes that cannot be supported using Cairo, in particular for nodes doing 3D operations, this function may fail.

cr

cairo context to draw to

pub fn bounds(&self) -> Rect

Retrieves the boundaries of the @self.

The node will not draw outside of its boundaries.

Returns

bounds

return location for the boundaries

pub fn node_type(&self) -> RenderNodeType

Returns the type of the @self.

Returns

the type of the RenderNode

pub fn serialize(&self) -> Bytes

Serializes the @self for later deserialization via gsk_render_node_deserialize(). No guarantees are made about the format used other than that the same version of GTK will be able to deserialize the result of a call to gsk_render_node_serialize() and gsk_render_node_deserialize() will correctly reject files it cannot open that were created with previous versions of GTK.

The intended use of this functions is testing, benchmarking and debugging. The format is not meant as a permanent storage format.

Returns

a GBytes representing the node.

pub fn write_to_file(&self, filename: impl AsRef<Path>) -> Result<(), Error>

This function is equivalent to calling [RenderNodeExt::serialize()][crate::prelude::RenderNodeExt::serialize()] followed by file_set_contents().

See those two functions for details on the arguments.

It is mostly intended for use inside a debugger to quickly dump a render node to a file for later inspection.

filename

the file to save it to.

Returns

true if saving was successful

impl RenderNode

pub fn deserialize(bytes: &Bytes) -> Option<Self>

Loads data previously created via [RenderNodeExt::serialize()][crate::prelude::RenderNodeExt::serialize()].

For a discussion of the supported format, see that function.

bytes

the bytes containing the data

error_func

Callback on parsing errors

Returns

a new RenderNode

pub fn deserialize_with_error_func<P: FnMut(&ParseLocation, &ParseLocation, &Error)>(
bytes: &Bytes,
error_func: P
) -> Option<Self>

pub fn downcast<T: IsRenderNode>(self) -> Result<T, Self>

pub fn downcast_ref<T: IsRenderNode>(&self) -> Option<&T>