Struct gsk4::ShaderArgsBuilder

pub struct ShaderArgsBuilder(_);

An object to build the uniforms data for a GLShader.

Implementations

impl ShaderArgsBuilder

pub fn new(
shader: &GLShader,
initial_values: Option<&Bytes>
) -> ShaderArgsBuilder

Allocates a builder that can be used to construct a new uniform data chunk.

shader

a GLShader

initial_values

optional GBytes with initial values

Returns

The newly allocated builder, free with Gsk::ShaderArgsBuilder::unref()``

pub fn free_to_args(&self) -> Option<Bytes>

Creates a new GBytes args from the current state of the given self, and frees the self instance.

Any uniforms of the shader that have not been explicitly set on the self are zero-initialized.

Returns

the newly allocated buffer with all the args added to self

pub fn set_bool(&self, idx: i32, value: bool)

Sets the value of the uniform idx.

The uniform must be of bool type.

idx

index of the uniform

value

value to set the uniform to

pub fn set_float(&self, idx: i32, value: f32)

Sets the value of the uniform idx.

The uniform must be of float type.

idx

index of the uniform

value

value to set the uniform to

pub fn set_int(&self, idx: i32, value: i32)

Sets the value of the uniform idx.

The uniform must be of int type.

idx

index of the uniform

value

value to set the uniform to

pub fn set_uint(&self, idx: i32, value: u32)

Sets the value of the uniform idx.

The uniform must be of uint type.

idx

index of the uniform

value

value to set the uniform to

pub fn set_vec2(&self, idx: i32, value: &Vec2)

Sets the value of the uniform idx.

The uniform must be of vec2 type.

idx

index of the uniform

value

value to set the uniform too

pub fn set_vec3(&self, idx: i32, value: &Vec3)

Sets the value of the uniform idx.

The uniform must be of vec3 type.

idx

index of the uniform

value

value to set the uniform too

pub fn set_vec4(&self, idx: i32, value: &Vec4)

Sets the value of the uniform idx.

The uniform must be of vec4 type.

idx

index of the uniform

value

value to set the uniform too

pub fn to_args(&self) -> Option<Bytes>

Creates a new GBytes args from the current state of the given self.

Any uniforms of the shader that have not been explicitly set on the self are zero-initialized.

The given ShaderArgsBuilder is reset once this function returns; you cannot call this function multiple times on the same self instance.

This function is intended primarily for bindings. C code should use free_to_args().

Returns

the newly allocated buffer with all the args added to self