Struct gsk4::TextNode

#[repr(transparent)]
pub struct TextNode { /* private fields */ }

A render node drawing a set of glyphs.

Implementations

impl TextNode

pub fn as_ptr(&self) -> *mut GskTextNode

Return the inner pointer to the underlying C value.

pub unsafe fn from_glib_ptr_borrow<'a>( ptr: *const *const GskTextNode ) -> &'a Self

Borrows the underlying C value.

impl TextNode

pub fn new( font: &impl IsA<Font>, glyphs: &GlyphString, color: &RGBA, offset: &Point ) -> Option<TextNode>

Creates a render node that renders the given glyphs.

Note that @color may not be used if the font contains color glyphs.

font

the pango::Font containing the glyphs

glyphs

the pango::GlyphString to render

color

the foreground color to render with

offset

offset of the baseline

Returns

a new RenderNode

pub fn color(&self) -> RGBA

Retrieves the color used by the text @self.

Returns

the text color

pub fn font(&self) -> Font

Returns the font used by the text @self.

Returns

the font

pub fn glyphs(&self) -> Vec<GlyphInfo>

Retrieves the glyph information in the @self.

Returns

the glyph information

pub fn num_glyphs(&self) -> u32

Retrieves the number of glyphs in the text node.

Returns

the number of glyphs

pub fn offset(&self) -> Point

Retrieves the offset applied to the text.

Returns

a point with the horizontal and vertical offsets

pub fn has_color_glyphs(&self) -> bool

Available on crate feature v4_2 only.

Checks whether the text @self has color glyphs.

Returns

true if the text node has color glyphs