Trait gsk4::RendererExt

pub trait RendererExt: 'static {
    pub fn get_surface(&self) -> Option<Surface>;

    pub fn is_realized(&self) -> bool;

    pub fn realize(&self, surface: &Surface) -> Result<(), Error>;

    pub fn render<P: IsA<RenderNode>>(&self, root: &P, region: Option<&Region>);

    pub fn render_texture<P: IsA<RenderNode>>(
        &self, 
        root: &P, 
        viewport: Option<&Rect>
    ) -> Option<Texture>;

    pub fn unrealize(&self);

    pub fn get_property_realized(&self) -> bool;

    pub fn connect_property_realized_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    pub fn connect_property_surface_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;
}

Trait containing all Renderer methods.

Implementors

BroadwayRenderer, CairoRenderer, GLRenderer, Renderer, VulkanRenderer

Required methods

pub fn get_surface(&self) -> Option<Surface>

Retrieves the gdk::Surface set using RendererExt::realize. If the renderer has not been realized yet, None will be returned.

Returns

a gdk::Surface

pub fn is_realized(&self) -> bool

Checks whether the self is realized or not.

Returns

true if the Renderer was realized, and false otherwise

pub fn realize(&self, surface: &Surface) -> Result<(), Error>

Creates the resources needed by the self to render the scene graph.

surface

the gdk::Surface renderer will be used on

pub fn render<P: IsA<RenderNode>>(&self, root: &P, region: Option<&Region>)

Renders the scene graph, described by a tree of RenderNode instances, ensuring that the given region gets redrawn.

Renderers must ensure that changes of the contents given by the root node as well as the area given by region are redrawn. They are however free to not redraw any pixel outside of region if they can guarantee that it didn't change.

The self will acquire a reference on the RenderNode tree while the rendering is in progress.

root

a RenderNode

region

the cairo::Region that must be redrawn or None for the whole window

pub fn render_texture<P: IsA<RenderNode>>(
&self,
root: &P,
viewport: Option<&Rect>
) -> Option<Texture>

Renders the scene graph, described by a tree of RenderNode instances, to a gdk::Texture.

The self will acquire a reference on the RenderNode tree while the rendering is in progress.

If you want to apply any transformations to root, you should put it into a transform node and pass that node instead.

root

a RenderNode

viewport

the section to draw or None to use root's bounds

Returns

a gdk::Texture with the rendered contents of root.

pub fn unrealize(&self)

Releases all the resources created by RendererExt::realize.

pub fn get_property_realized(&self) -> bool

pub fn connect_property_realized_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

pub fn connect_property_surface_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

Loading content...