Crate gsk4

Modules

prelude
Traits intended for blanket imports.

Structs

BlendNode

BlendNodeClass

BlurNode

BlurNodeClass

BorderNode

BorderNodeClass

CairoNode

CairoNodeClass

CairoRenderer

CairoRendererClass

ClipNode

ClipNodeClass

ColorMatrixNode

ColorMatrixNodeClass

ColorNode

ColorNodeClass

ColorStop

ContainerNode

ContainerNodeClass

CrossFadeNode

CrossFadeNodeClass

DebugNode

DebugNodeClass

GLRenderer

GLRendererClass

InsetShadowNode

InsetShadowNodeClass

LinearGradientNode

LinearGradientNodeClass

OpacityNode

OpacityNodeClass

OutsetShadowNode

OutsetShadowNodeClass

RenderNode

RenderNodeClass

Renderer

RendererClass

RepeatNode

RepeatNodeClass

RepeatingLinearGradientNode

RepeatingLinearGradientNodeClass

RoundedClipNode

RoundedClipNodeClass

RoundedRect

Shadow

ShadowNode

ShadowNodeClass

TextNode

TextNodeClass

TextureNode

TextureNodeClass

Transform

TransformNode

TransformNodeClass

Enums

BlendMode
The blend modes available for render nodes.

Corner
The corner indices used by RoundedRect.

RenderNodeType
The type of a node determines what the node is rendering.

ScalingFilter
The filters used when scaling texture data.

SerializationError
Errors that can happen during (de)serialization.

TransformCategory
The categories of matrices relevant for GSK and GTK. Note that any category includes matrices of all later categories. So if you want to for example check if a matrix is a 2D matrix, category >= GSK_TRANSFORM_CATEGORY_2D is the way to do this.

Constants

NONE_RENDERER

NONE_RENDER_NODE

Traits

BorderNodeManualExt

RenderNodeExt
Trait containing all RenderNode methods.

RendererExt
Trait containing all Renderer methods.