Enum gsk4::BlendMode

#[non_exhaustive]
pub enum BlendMode {

    Default,
    Multiply,
    Screen,
    Overlay,
    Darken,
    Lighten,
    ColorDodge,
    ColorBurn,
    HardLight,
    SoftLight,
    Difference,
    Exclusion,
    Color,
    Hue,
    Saturation,
    Luminosity,
    // some variants omitted

}

The blend modes available for render nodes.

The implementation of each blend mode is deferred to the rendering pipeline.

See https://www.w3.org/TR/compositing-1/#blending for more information on blending and blend modes.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Default

The default blend mode, which specifies no blending

Multiply

The source color is multiplied by the destination and replaces the destination

Screen

Multiplies the complements of the destination and source color values, then complements the result.

Overlay

Multiplies or screens the colors, depending on the destination color value. This is the inverse of hard-list

Darken

Selects the darker of the destination and source colors

Lighten

Selects the lighter of the destination and source colors

ColorDodge

Brightens the destination color to reflect the source color

ColorBurn

Darkens the destination color to reflect the source color

HardLight

Multiplies or screens the colors, depending on the source color value

SoftLight

Darkens or lightens the colors, depending on the source color value

Difference

Subtracts the darker of the two constituent colors from the lighter color

Exclusion

Produces an effect similar to that of the difference mode but lower in contrast

Color

Creates a color with the hue and saturation of the source color and the luminosity of the destination color

Hue

Creates a color with the hue of the source color and the saturation and luminosity of the destination color

Saturation

Creates a color with the saturation of the source color and the hue and luminosity of the destination color

Luminosity

Creates a color with the luminosity of the source color and the hue and saturation of the destination color