Enum gsk4::RenderNodeType

#[non_exhaustive]
pub enum RenderNodeType {

    NotARenderNode,
    ContainerNode,
    CairoNode,
    ColorNode,
    LinearGradientNode,
    RepeatingLinearGradientNode,
    RadialGradientNode,
    RepeatingRadialGradientNode,
    ConicGradientNode,
    BorderNode,
    TextureNode,
    InsetShadowNode,
    OutsetShadowNode,
    TransformNode,
    OpacityNode,
    ColorMatrixNode,
    RepeatNode,
    ClipNode,
    RoundedClipNode,
    ShadowNode,
    BlendNode,
    CrossFadeNode,
    TextNode,
    BlurNode,
    DebugNode,
    GlShaderNode,
    TextureScaleNode,
    MaskNode,
    FillNode,
    StrokeNode,
    SubsurfaceNode,

}

The type of a node determines what the node is rendering.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

NotARenderNode

Error type. No node will ever have this type.

ContainerNode

A node containing a stack of children

CairoNode

A node drawing a cairo::Surface

ColorNode

A node drawing a single color rectangle

LinearGradientNode

A node drawing a linear gradient

RepeatingLinearGradientNode

A node drawing a repeating linear gradient

RadialGradientNode

A node drawing a radial gradient

RepeatingRadialGradientNode

A node drawing a repeating radial gradient

ConicGradientNode

A node drawing a conic gradient

BorderNode

A node stroking a border around an area

TextureNode

A node drawing a gdk::Texture

InsetShadowNode

A node drawing an inset shadow

OutsetShadowNode

A node drawing an outset shadow

TransformNode

A node that renders its child after applying a matrix transform

OpacityNode

A node that changes the opacity of its child

ColorMatrixNode

A node that applies a color matrix to every pixel

RepeatNode

A node that repeats the child’s contents

ClipNode

A node that clips its child to a rectangular area

RoundedClipNode

A node that clips its child to a rounded rectangle

ShadowNode

A node that draws a shadow below its child

BlendNode

A node that blends two children together

CrossFadeNode

A node that cross-fades between two children

TextNode

A node containing a glyph string

BlurNode

A node that applies a blur

DebugNode

Debug information that does not affect the rendering

GlShaderNode

A node that uses OpenGL fragment shaders to render

TextureScaleNode

Available on crate feature v4_10 only.

A node drawing a gdk::Texture scaled and filtered.

MaskNode

Available on crate feature v4_10 only.

A node that masks one child with another.

FillNode

Available on crate feature v4_14 only.

A node that fills a path.

StrokeNode

Available on crate feature v4_14 only.

A node that strokes a path.

SubsurfaceNode

Available on crate feature v4_14 only.

A node that possibly redirects part of the scene graph to a subsurface.