gsk4

Struct ComponentTransferNode

pub struct ComponentTransferNode { /* private fields */ }

Available on crate feature v4_20 only.

A render node for applying a ComponentTransfer for each color component of the child node.

GLib type: Shared boxed type with reference counted clone semantics.

Implementations

impl ComponentTransferNode

pub fn as_ptr(&self) -> *mut GskComponentTransferNode

Return the inner pointer to the underlying C value.

pub unsafe fn from_glib_ptr_borrow(ptr: &*mut GskComponentTransferNode) -> &Self

Borrows the underlying C value.

impl ComponentTransferNode

pub fn new( child: impl AsRef<RenderNode>, r: &ComponentTransfer, g: &ComponentTransfer, b: &ComponentTransfer, a: &ComponentTransfer, ) -> ComponentTransferNode

Creates a render node that will apply component transfers to a child node.

child

The child to apply component transfers to

r

transfer for the red component

g

transfer for the green component

b

transfer for the blue component

a

transfer for the alpha component

Returns

A new RenderNode

pub fn child(&self) -> RenderNode

Gets the child node that is getting drawn by the given @self.

Returns

the child RenderNode

pub fn transfer(&self, component: u32) -> ComponentTransfer

Gets the component transfer for one of the components.

component

a value between 0 and 3 to indicate the red, green, blue or alpha component

Returns

the ComponentTransfer