Struct gsk4::GLShader

pub struct GLShader(_);

A GLShader is a snippet of GLSL that is meant to run in the fragment shader of the rendering pipeline.

A fragment shader gets the coordinates being rendered as input and produces the pixel values for that particular pixel. Additionally, the shader can declare a set of other input arguments, called uniforms (as they are uniform over all the calls to your shader in each instance of use). A shader can also receive up to 4 textures that it can use as input when producing the pixel data.

GLShader is usually used with gtk_snapshot_push_gl_shader() to produce a GLShaderNode in the rendering hierarchy, and then its input textures are constructed by rendering the child nodes to textures before rendering the shader node itself. (You can pass texture nodes as children if you want to directly use a texture as input).

The actual shader code is GLSL code that gets combined with some other code into the fragment shader. Since the exact capabilities of the GPU driver differs between different OpenGL drivers and hardware, GTK adds some defines that you can use to ensure your GLSL code runs on as many drivers as it can.

If the OpenGL driver is GLES, then the shader language version is set to 100, and GSK_GLES will be defined in the shader.

Otherwise, if the OpenGL driver does not support the 3.2 core profile, then the shader will run with language version 110 for GL2 and 130 for GL3, and GSK_LEGACY will be defined in the shader.

If the OpenGL driver supports the 3.2 code profile, it will be used, the shader language version is set to 150, and GSK_GL3 will be defined in the shader.

The main function the shader must implement is:

⚠️ The following code is in glsl ⚠️

 void mainImage(out vec4 fragColor,
                in vec2 fragCoord,
                in vec2 resolution,
                in vec2 uv)

Where the input fragCoord is the coordinate of the pixel we’re currently rendering, relative to the boundary rectangle that was specified in the GLShaderNode, and resolution is the width and height of that rectangle. This is in the typical GTK coordinate system with the origin in the top left. uv contains the u and v coordinates that can be used to index a texture at the corresponding point. These coordinates are in the [0..1]x[0..1] region, with 0, 0 being in the lower left corder (which is typical for OpenGL).

The output fragColor should be a RGBA color (with premultiplied alpha) that will be used as the output for the specified pixel location. Note that this output will be automatically clipped to the clip region of the glshader node.

In addition to the function arguments the shader can define up to 4 uniforms for textures which must be called u_textureN (i.e. u_texture1 to u_texture4) as well as any custom uniforms you want of types int, uint, bool, float, vec2, vec3 or vec4.

All textures sources contain premultiplied alpha colors, but if some there are outer sources of colors there is a gsk_premultiply() helper to compute premultiplication when needed.

Note that GTK parses the uniform declarations, so each uniform has to be on a line by itself with no other code, like so:

⚠️ The following code is in glsl ⚠️

uniform float u_time;
uniform vec3 u_color;
uniform sampler2D u_texture1;
uniform sampler2D u_texture2;

GTK uses the the “gsk” namespace in the symbols it uses in the shader, so your code should not use any symbols with the prefix gsk or GSK. There are some helper functions declared that you can use:

⚠️ The following code is in glsl ⚠️

vec4 GskTexture(sampler2D sampler, vec2 texCoords);

This samples a texture (e.g. u_texture1) at the specified coordinates, and containes some helper ifdefs to ensure that it works on all OpenGL versions.

You can compile the shader yourself using compile(), otherwise the GSK renderer will do it when it handling the glshader node. If errors occurs, the returned error will include the glsl sources, so you can see what GSK was passing to the compiler. You can also set GSK_DEBUG=shaders in the environment to see the sources and other relevant information about all shaders that GSK is handling.

An example shader

⚠️ The following code is in glsl ⚠️

uniform float position;
uniform sampler2D u_texture1;
uniform sampler2D u_texture2;

void mainImage(out vec4 fragColor,
               in vec2 fragCoord,
               in vec2 resolution,
               in vec2 uv) {
  vec4 source1 = GskTexture(u_texture1, uv);
  vec4 source2 = GskTexture(u_texture2, uv);

  fragColor = position * source1 + (1.0 - position) * source2;
}

Implementations

impl GLShader

pub fn from_bytes(sourcecode: &Bytes) -> GLShader

Creates a GLShader that will render pixels using the specified code.

sourcecode

GLSL sourcecode for the shader, as a GBytes

Returns

A new GLShader

pub fn from_resource(resource_path: &str) -> GLShader

Creates a GLShader that will render pixels using the specified code.

resource_path

path to a resource that contains the GLSL sourcecode for the shader

Returns

A new GLShader

pub fn builder() -> GLShaderBuilder

Creates a new builder-pattern struct instance to construct GLShader objects.

This method returns an instance of GLShaderBuilder which can be used to create GLShader objects.

pub fn compile<P: IsA<Renderer>>(&self, renderer: &P) -> Result<(), Error>

Tries to compile the self for the given renderer.

If there is a problem, this function returns false and reports an error. You should use this function before relying on the shader for rendering and use a fallback with a simpler shader or without shaders if it fails.

Note that this will modify the rendering state (for example change the current GL context) and requires the renderer to be set up. This means that the widget has to be realized. Commonly you want to call this from the realize signal of a widget, or during widget snapshot.

renderer

a Renderer

Returns

true on success, false if an error occurred

pub fn find_uniform_by_name(&self, name: &str) -> i32

Looks for a uniform by the name name, and returns the index of the uniform, or -1 if it was not found.

name

uniform name

Returns

The index of the uniform, or -1

pub fn arg_bool(&self, args: &Bytes, idx: i32) -> bool

Gets the value of the uniform idx in the args block.

The uniform must be of bool type.

args

uniform arguments

idx

index of the uniform

Returns

The value

pub fn arg_float(&self, args: &Bytes, idx: i32) -> f32

Gets the value of the uniform idx in the args block.

The uniform must be of float type.

args

uniform arguments

idx

index of the uniform

Returns

The value

pub fn arg_int(&self, args: &Bytes, idx: i32) -> i32

Gets the value of the uniform idx in the args block.

The uniform must be of int type.

args

uniform arguments

idx

index of the uniform

Returns

The value

pub fn arg_uint(&self, args: &Bytes, idx: i32) -> u32

Gets the value of the uniform idx in the args block.

The uniform must be of uint type.

args

uniform arguments

idx

index of the uniform

Returns

The value

pub fn args_size(&self) -> usize

Get the size of the data block used to specify arguments for this shader.

Returns

The size of the data block

pub fn n_textures(&self) -> i32

Returns the number of textures that the shader requires.

This can be used to check that the a passed shader works in your usecase. It is determined by looking at the highest u_textureN value that the shader defines.

Returns

The number of texture inputs required by self

pub fn n_uniforms(&self) -> i32

Get the number of declared uniforms for this shader.

Returns

The number of declared uniforms

pub fn resource(&self) -> Option<GString>

Gets the resource path for the GLSL sourcecode being used to render this shader.

Returns

The resource path for the shader

pub fn source(&self) -> Option<Bytes>

Gets the GLSL sourcecode being used to render this shader.

Returns

The source code for the shader

pub fn uniform_name(&self, idx: i32) -> Option<GString>

Get the name of the declared uniform for this shader at index idx.

idx

index of the uniform

Returns

The name of the declared uniform

pub fn uniform_offset(&self, idx: i32) -> i32

Get the offset into the data block where data for this uniforms is stored.

idx

index of the uniform

Returns

The data offset

pub fn uniform_type(&self, idx: i32) -> GLUniformType

Get the type of the declared uniform for this shader at index idx.

idx

index of the uniform

Returns

The type of the declared uniform

impl GLShader

pub fn arg_vec2(&self, args: &Bytes, idx: i32) -> Vec2

Gets the value of the uniform idx in the args block.

The uniform must be of vec2 type.

args

uniform arguments

idx

index of the uniform

out_value

location to store the uniform value in

pub fn arg_vec3(&self, args: &Bytes, idx: i32) -> Vec3

Gets the value of the uniform idx in the args block.

The uniform must be of vec3 type.

args

uniform arguments

idx

index of the uniform

out_value

location to store the uniform value in

pub fn arg_vec4(&self, args: &Bytes, idx: i32) -> Vec4

Gets the value of the uniform idx in the args block.

The uniform must be of vec4 type.

args

uniform arguments

idx

index of the uniform

out_value

location to store set the uniform value in