Struct gsk4::RenderNode

pub struct RenderNode(_);

A node in the render tree.

This is an Abstract Base Class, you cannot instantiate it.

Implements

RenderNodeExt

Implementations

impl RenderNode

pub fn deserialize(bytes: &Bytes) -> Option<RenderNode>

Loads data previously created via RenderNodeExt::serialize. For a discussion of the supported format, see that function.

bytes

the bytes containing the data

error_func

Callback on parsing errors or None

user_data

user_data for error_func

Returns

a new RenderNode or None on error.

pub fn deserialize_with_error_func<P: FnMut(&ParseLocation, &ParseLocation, &Error)>(
bytes: &Bytes,
error_func: P
) -> Option<RenderNode>

pub fn downcast<T: IsRenderNode>(self) -> Result<T, RenderNode>

pub fn downcast_ref<T: IsRenderNode>(&self) -> Option<&T>

pub fn draw(&self, cr: &Context)

pub fn get_bounds(&self) -> Rect

pub fn get_node_type(&self) -> RenderNodeType

pub fn serialize(&self) -> Bytes

pub fn write_to_file<P: AsRef<Path>>(&self, filename: P) -> Result<(), Error>