gsk4

Struct RenderNode

pub struct RenderNode { /* private fields */ }

The basic block in a scene graph to be rendered using Renderer.

Each node has a parent, except the top-level node; each node may have children nodes.

Each node has an associated drawing surface, which has the size of the rectangle set when creating it.

Render nodes are meant to be transient; once they have been associated to a Renderer it’s safe to release any reference you have on them. All RenderNodes are immutable, you can only specify their properties during construction.

This is an Abstract Base Class, you cannot instantiate it.

GLib type: Shared boxed type with reference counted clone semantics.

Implementations

impl RenderNode

pub fn as_ptr(&self) -> *mut GskRenderNode

Return the inner pointer to the underlying C value.

pub unsafe fn from_glib_ptr_borrow(ptr: &*mut GskRenderNode) -> &Self

Borrows the underlying C value.

impl RenderNode

pub const NONE: Option<&'static RenderNode> = None

pub fn draw(&self, cr: &Context)

Draws the contents of a render node on a cairo context.

Typically, you’ll use this function to implement fallback rendering of render nodes on an intermediate Cairo context, instead of using the drawing context associated to a gdk::Surface’s rendering buffer.

For advanced nodes that cannot be supported using Cairo, in particular for nodes doing 3D operations, this function may fail.

cr

cairo context to draw to

pub fn bounds(&self) -> Rect

Retrieves the boundaries of the @self.

The node will not draw outside of its boundaries.

Returns

bounds

return location for the boundaries

pub fn node_type(&self) -> RenderNodeType

Returns the type of the render node.

Returns

the type of @self

pub fn opaque_rect(&self) -> Option<Rect>

Available on crate feature v4_16 only.

Gets an opaque rectangle inside the node that GTK can determine to be fully opaque.

There is no guarantee that this is indeed the largest opaque rectangle or that regions outside the rectangle are not opaque. This function is a best effort with that goal.

The rectangle will be fully contained in the bounds of the node.

Returns

true if part or all of the rendernode is opaque, false if no opaque region could be found.

out_opaque

return location for the opaque rect

pub fn serialize(&self) -> Bytes

Serializes the @self for later deserialization via gsk_render_node_deserialize(). No guarantees are made about the format used other than that the same version of GTK will be able to deserialize the result of a call to gsk_render_node_serialize() and gsk_render_node_deserialize() will correctly reject files it cannot open that were created with previous versions of GTK.

The intended use of this functions is testing, benchmarking and debugging. The format is not meant as a permanent storage format.

Returns

a GBytes representing the node.

pub fn write_to_file(&self, filename: impl AsRef<Path>) -> Result<(), Error>

This function is equivalent to calling serialize() followed by file_set_contents().

See those two functions for details on the arguments.

It is mostly intended for use inside a debugger to quickly dump a render node to a file for later inspection.

filename

the file to save it to

Returns

true if saving was successful

impl RenderNode

pub fn is<T: IsRenderNode>(&self) -> bool

pub fn type_(&self) -> Type

pub fn deserialize(bytes: &Bytes) -> Option<Self>

Loads data previously created via serialize().

For a discussion of the supported format, see that function.

bytes

the bytes containing the data

error_func

callback on parsing errors

Returns

a new render node

pub fn deserialize_with_error_func<P: FnMut(&ParseLocation, &ParseLocation, &Error)>( bytes: &Bytes, error_func: P, ) -> Option<Self>

pub fn downcast<T: IsRenderNode>(self) -> Result<T, Self>

pub fn downcast_ref<T: IsRenderNode>(&self) -> Option<&T>