Struct gsk4::Stroke

pub struct Stroke { /* private fields */ }

Available on crate feature v4_14 only.

A Stroke struct collects the parameters that influence the operation of stroking a path.

Implementations

impl Stroke

pub fn as_ptr(&self) -> *mut GskStroke

Return the inner pointer to the underlying C value.

pub unsafe fn from_glib_ptr_borrow<'a>(ptr: *const *const GskStroke) -> &'a Self

Borrows the underlying C value.

pub unsafe fn from_glib_ptr_borrow_mut<'a>( ptr: *mut *mut GskStroke ) -> &'a mut Self

Borrows the underlying C value mutably.

impl Stroke

pub fn new(line_width: f32) -> Stroke

Creates a new Stroke with the given @line_width.

line_width

line width of the stroke. Must be > 0

Returns

a new Stroke

pub fn dash(&self) -> Vec<f32>

Gets the dash array in use or NULL if dashing is disabled.

Returns

The dash array or NULL if the dash array is empty.

pub fn dash_offset(&self) -> f32

Returns the dash_offset of a Stroke.

pub fn line_cap(&self) -> LineCap

Gets the line cap used.

See LineCap for details.

Returns

The line cap

pub fn line_join(&self) -> LineJoin

Gets the line join used.

See LineJoin for details.

Returns

The line join

pub fn line_width(&self) -> f32

Gets the line width used.

Returns

The line width

pub fn miter_limit(&self) -> f32

Returns the miter limit of a Stroke.

pub fn set_dash(&self, dash: &[f32])

Sets the dash pattern to use by this stroke.

A dash pattern is specified by an array of alternating non-negative values. Each value provides the length of alternate “on” and “off” portions of the stroke.

Each “on” segment will have caps applied as if the segment were a separate contour. In particular, it is valid to use an “on” length of 0 with GSK_LINE_CAP_ROUND or GSK_LINE_CAP_SQUARE to draw dots or squares along a path.

If @n_dash is 0, if all elements in @dash are 0, or if there are negative values in @dash, then dashing is disabled.

If @n_dash is 1, an alternating “on” and “off” pattern with the single dash length provided is assumed.

If @n_dash is uneven, the dash array will be used with the first element in @dash defining an “on” or “off” in alternating passes through the array.

You can specify a starting offset into the dash with set_dash_offset().

dash

the array of dashes

pub fn set_dash_offset(&self, offset: f32)

Sets the offset into the dash pattern where dashing should begin.

This is an offset into the length of the path, not an index into the array values of the dash array.

See set_dash() for more details on dashing.

offset

offset into the dash pattern

pub fn set_line_cap(&self, line_cap: LineCap)

Sets the line cap to be used when stroking.

See LineCap for details.

line_cap

the LineCap

pub fn set_line_join(&self, line_join: LineJoin)

Sets the line join to be used when stroking.

See LineJoin for details.

line_join

The line join to use

pub fn set_line_width(&self, line_width: f32)

Sets the line width to be used when stroking.

The line width must be > 0.

line_width

width of the line in pixels

pub fn set_miter_limit(&self, limit: f32)

Sets the limit for the distance from the corner where sharp turns of joins get cut off.

The miter limit is in units of line width and must be non-negative.

For joins of type GSK_LINE_JOIN_MITER that exceed the miter limit, the join gets rendered as if it was of type GSK_LINE_JOIN_BEVEL.

limit

the miter limit

pub fn to_cairo(&self, cr: &Context)

A helper function that sets the stroke parameters of @cr from the values found in @self.

cr

the cairo context to configure

impl Stroke

pub fn builder(line_width: f32) -> StrokeBuilder

Creates a new builder-pattern struct instance to construct a Stroke.

This method returns an instance of StrokeBuilder which can be used to create a Stroke.