Struct gsk4::Transform

pub struct Transform(_);

Transform is an object to describe transform matrices.

Unlike graphene_matrix_t, Transform retains the steps in how a transform was constructed, and allows inspecting them. It is modeled after the way CSS describes transforms.

Transform objects are immutable and cannot be changed after creation. This means code can safely expose them as properties of objects without having to worry about others changing them.

Implementations

impl Transform

pub fn new() -> Transform

pub fn category(&self) -> TransformCategory

Returns the category this transform belongs to.

Returns

The category of the transform

pub fn invert(&self) -> Option<Transform>

Inverts the given transform.

If self is not invertible, None is returned. Note that inverting None also returns None, which is the correct inverse of None. If you need to differentiate between those cases, you should check self is not None before calling this function.

Returns

The inverted transform

pub fn matrix(&self, matrix: &Matrix) -> Option<Transform>

Multiplies self with the given matrix.

matrix

the matrix to multiply self with

Returns

The new transform

pub fn perspective(&self, depth: f32) -> Option<Transform>

Applies a perspective projection transform.

This transform scales points in X and Y based on their Z value, scaling points with positive Z values away from the origin, and those with negative Z values towards the origin. Points on the z=0 plane are unchanged.

depth

distance of the z=0 plane. Lower values give a more flattened pyramid and therefore a more pronounced perspective effect.

Returns

The new transform

pub fn rotate(&self, angle: f32) -> Option<Transform>

Rotates self angle degrees in 2D - or in 3D-speak, around the z axis.

angle

the rotation angle, in degrees (clockwise)

Returns

The new transform

pub fn rotate_3d(&self, angle: f32, axis: &Vec3) -> Option<Transform>

Rotates self angle degrees around axis.

For a rotation in 2D space, use rotate()

angle

the rotation angle, in degrees (clockwise)

axis

The rotation axis

Returns

The new transform

pub fn scale(&self, factor_x: f32, factor_y: f32) -> Option<Transform>

Scales self in 2-dimensional space by the given factors.

Use scale_3d() to scale in all 3 dimensions.

factor_x

scaling factor on the X axis

factor_y

scaling factor on the Y axis

Returns

The new transform

pub fn scale_3d(
&self,
factor_x: f32,
factor_y: f32,
factor_z: f32
) -> Option<Transform>

Scales self by the given factors.

factor_x

scaling factor on the X axis

factor_y

scaling factor on the Y axis

factor_z

scaling factor on the Z axis

Returns

The new transform

pub fn to_2d(&self) -> (f32, f32, f32, f32, f32, f32)

Converts a Transform to a 2D transformation matrix.

self must be a 2D transformation. If you are not sure, use category() >= TransformCategory::_2d to check.

The returned values have the following layout:

  | xx yx |   |  a  b  0 |
  | xy yy | = |  c  d  0 |
  | dx dy |   | tx ty  1 |

This function can be used to convert between a Transform and a matrix type from other 2D drawing libraries, in particular Cairo.

Returns

out_xx

return location for the xx member

out_yx

return location for the yx member

out_xy

return location for the xy member

out_yy

return location for the yy member

out_dx

return location for the x0 member

out_dy

return location for the y0 member

pub fn to_affine(&self) -> (f32, f32, f32, f32)

Converts a Transform to 2D affine transformation factors.

self must be a 2D transformation. If you are not sure, use

category() >= TransformCategory::_2dAffine

to check.

Returns

out_scale_x

return location for the scale factor in the x direction

out_scale_y

return location for the scale factor in the y direction

out_dx

return location for the translation in the x direction

out_dy

return location for the translation in the y direction

pub fn to_matrix(&self) -> Matrix

Computes the actual value of self and stores it in out_matrix.

The previous value of out_matrix will be ignored.

Returns

out_matrix

The matrix to set

pub fn to_str(&self) -> GString

pub fn to_translate(&self) -> (f32, f32)

Converts a Transform to a translation operation.

self must be a 2D transformation. If you are not sure, use

category() >= TransformCategory::_2dTranslate

to check.

Returns

out_dx

return location for the translation in the x direction

out_dy

return location for the translation in the y direction

pub fn transform(&self, other: Option<&Transform>) -> Option<Transform>

Applies all the operations from other to self.

other

Transform to apply

Returns

The new transform

pub fn transform_bounds(&self, rect: &Rect) -> Rect

Transforms a graphene_rect_t using the given transform self.

The result is the bounding box containing the coplanar quad.

rect

a graphene_rect_t

Returns

out_rect

return location for the bounds of the transformed rectangle

pub fn transform_point(&self, point: &Point) -> Point

Transforms a graphene_point_t using the given transform self.

point

a graphene_point_t

Returns

out_point

return location for the transformed point

pub fn translate(&self, point: &Point) -> Option<Transform>

Translates self in 2-dimensional space by point.

point

the point to translate the transform by

Returns

The new transform

pub fn translate_3d(&self, point: &Point3D) -> Option<Transform>

Translates self by point.

point

the point to translate the transform by

Returns

The new transform

pub fn parse(string: &str) -> Option<Transform>

Parses the given string into a transform and puts it in out_transform.

Strings printed via Gsk::Transform::to_string()`` can be read in again successfully using this function.

If string does not describe a valid transform, false is returned and None is put in out_transform.

string

the string to parse

Returns

true if string described a valid transform.

out_transform

The location to put the transform in